moved layouts to its own components folder
