--- src/components/LayoutExample/index.js
@@ -1,3 +1,0 @@
-import LayoutExample from './LayoutExample';
-
-export default LayoutExample;
